created send_query function
